Switch to official Android Compose screenshot testing with Roborazzi
Replaces Paparazzi with Roborazzi, the official Android-recommended screenshot
testing library, and adds PR-based snapshot update functionality.

## Major Changes

### Testing Framework Migration
- **Removed**: Paparazzi 1.3.4
- **Added**: Roborazzi 1.26.0 + Robolectric 4.13
- **Why**: Roborazzi is the recommended approach per
  https://developer.android.com/studio/preview/compose-screenshot-testing
- Uses official Android screenshot testing APIs
- Better Compose integration
- More reliable rendering

### PR-Based Snapshot Updates 🎉
Added `/record-snapshots` command support in PRs:
- Comment `/record-snapshots` or `/update-snapshots` on any PR
- GitHub Actions automatically records new snapshots
- Commits them directly to the PR branch
- No need to run tests locally to update snapshots!

### Test Suite Optimization
Consolidated from 6 test files to 3 focused test files:
- **ComponentSnapshotTests** (~25 tests): Individual UI components
- **ColorPickerSnapshotTests** (~10 tests): Full ColorPicker configurations
- **DialogAndConfigurationSnapshotTests** (~15 tests): Dialogs, preferences,
  device configs, themes, and RTL support

Total: **~50 comprehensive tests** (down from 110+ for better maintainability)

### Configuration Updates

**build.gradle.kts:**
- Added Roborazzi plugin
- Added Robolectric dependency
- Configured test options for native graphics mode
- Updated Gradle tasks:
  - `recordSnapshots` → uses `recordRoborazziDebug`
  - `verifySnapshots` → uses `verifyRoborazziDebug`
  - `compareSnapshots` → uses `compareRoborazziDebug`
  - Simplified task implementations

**Test API Changes:**
- Uses `@RunWith(RobolectricTestRunner::class)`
- Uses `@GraphicsMode(GraphicsMode.Mode.NATIVE)`
- Uses `@Config(sdk, qualifiers)` for device configuration
- Uses `createComposeRule()` instead of Paparazzi rule
- Uses `captureRoboImage()` to capture snapshots

**Snapshot Storage:**
- Old: `src/test/snapshots/`
- New: `src/test/resources/roborazzi/`
- Comparison images: `build/outputs/roborazzi/`

### GitHub Workflow Enhancements

**New Capabilities:**
1. **PR Comment Commands**:
   - Listens for `/record-snapshots` or `/update-snapshots` comments
   - Automatically records and commits snapshots
   - Posts results back to PR

2. **Improved Failure Reporting**:
   - Uploads comparison images showing exact differences
   - Highlights mismatched pixels
   - Provides clear instructions for fixing

3. **Multiple Fix Options**:
   - Option 1: Comment `/record-snapshots` (easiest!)
   - Option 2: Download artifacts and fix regression
   - Option 3: Update locally and push

**Permissions Added:**
- `contents: write` - For committing snapshots
- `pull-requests: write` - For commenting
- `issues: write` - For issue comments

### Documentation Updates

**library/SNAPSHOT_TESTING.md:**
- Updated for Roborazzi API
- Added PR comment command documentation
- Simplified to ~50 tests coverage
- Added configuration examples
- Removed Paparazzi references

**README.md:**
- Updated testing section with Roborazzi
- Added PR snapshot update feature
- Adjusted test counts (~50 tests)
- Added quick example of `/record-snapshots` command

**.gitignore:**
- Updated to ignore Roborazzi comparison artifacts:
  - `**/build/outputs/roborazzi/`
  - `**/*_compare.png`
  - `**/*_actual.png`
- Reference snapshots are still tracked in git

### Benefits

✅ **Official Android Support**: Uses recommended testing approach
✅ **Better Developer Experience**: Update snapshots from PR comments
✅ **Faster CI**: JVM-based, no emulator required
✅ **More Reliable**: Native graphics mode for accurate rendering
✅ **Better Maintainability**: Consolidated test suite (50 vs 110+ tests)
✅ **Improved Workflow**: No need to run tests locally to update snapshots
✅ **Clear Diff Images**: Roborazzi highlights exact pixel differences

### Migration Notes

For developers with existing checkouts:
1. Delete old snapshot directory: `rm -rf library/src/test/snapshots/`
2. Record new snapshots: `./gradlew :library:recordSnapshots`
3. Snapshots will be in: `library/src/test/resources/roborazzi/`

For PR reviews:
- If snapshots fail, comment `/record-snapshots` to update
- Review the comparison artifacts to see what changed
- Snapshots will be auto-committed to the PR

### Test Execution

```bash
# Record reference snapshots
./gradlew :library:recordSnapshots

# Verify snapshots
./gradlew :library:verifySnapshots

# Compare and generate diff images
./gradlew :library:compareSnapshots
```
