+++ build.gradle.kts
@@ -5,5 +5,5 @@ plugins {
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.kotlin.compose) apply false
     alias(libs.plugins.detekt) apply false
-    alias(libs.plugins.paparazzi) apply false
+    alias(libs.plugins.roborazzi) apply false
 }
