+++ library/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
     alias(libs.plugins.detekt)
-    alias(libs.plugins.paparazzi)
+    alias(libs.plugins.roborazzi)
 }
 
 android {
@@ -45,6 +45,16 @@ android {
         targetSdk = 36
     }
 
+    testOptions {
+        unitTests {
+            isIncludeAndroidResources = true
+            all {
+                it.systemProperty("robolectric.graphicsMode", "native")
+                it.systemProperty("roborazzi.test.record", System.getProperty("roborazzi.test.record"))
+            }
+        }
+    }
+
 }
 
 dependencies {
@@ -66,6 +76,11 @@ dependencies {
     testImplementation(libs.mockito.core)
     testImplementation(libs.mockito.kotlin)
     testImplementation(libs.kotlinx.coroutines.test)
+    testImplementation(libs.robolectric)
+    testImplementation(libs.roborazzi)
+    testImplementation(libs.roborazzi.compose)
+    testImplementation(libs.roborazzi.rule)
+    testImplementation(libs.androidx.compose.ui.test.junit4)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
     androidTestImplementation(platform(libs.androidx.compose.bom))
@@ -108,7 +123,7 @@ dependencies {
 tasks.register("recordSnapshots") {
     group = "snapshot"
     description = "Records reference snapshots for all snapshot tests"
-    dependsOn("recordPaparazziDebug")
+    dependsOn("recordRoborazziDebug")
 }
 
 /**
@@ -120,88 +135,19 @@ tasks.register("recordSnapshots") {
 tasks.register("verifySnapshots") {
     group = "snapshot"
     description = "Verifies snapshots match the reference images"
-    dependsOn("verifyPaparazziDebug")
+    dependsOn("verifyRoborazziDebug")
 }
 
 /**
- * Task to record comparison snapshots in a separate directory.
- * This creates a parallel set of snapshots for manual comparison.
- *
- * Usage: ./gradlew :library:recordCompareSnapshots
- */
-tasks.register("recordCompareSnapshots") {
-    group = "snapshot"
-    description = "Records comparison snapshots in a separate directory for manual comparison"
-    doLast {
-        // First record the snapshots
-        project.exec {
-            commandLine("./gradlew", ":library:recordPaparazziDebug")
-        }
-
-        // Copy snapshots to comparison directory
-        val snapshotsDir = file("src/test/snapshots")
-        val compareDir = file("src/test/snapshots-compare")
-
-        if (snapshotsDir.exists()) {
-            delete(compareDir)
-            copy {
-                from(snapshotsDir)
-                into(compareDir)
-            }
-            println("Comparison snapshots recorded to: $compareDir")
-        }
-    }
-}
-
-/**
- * Task to compare snapshots between reference and comparison directories.
- * Reports all differences without stopping at the first failure.
+ * Task to compare snapshots (generates comparison report).
+ * This task runs the tests in compare mode and generates a report of differences.
  *
  * Usage: ./gradlew :library:compareSnapshots
  */
 tasks.register("compareSnapshots") {
     group = "snapshot"
-    description = "Compares all snapshots and reports all failures"
-    doLast {
-        val snapshotsDir = file("src/test/snapshots")
-        val compareDir = file("src/test/snapshots-compare")
-
-        if (!snapshotsDir.exists() || !compareDir.exists()) {
-            throw GradleException("Missing snapshot directories. Run recordSnapshots and recordCompareSnapshots first.")
-        }
-
-        val failures = mutableListOf<String>()
-
-        snapshotsDir.walkTopDown()
-            .filter { it.isFile && it.extension == "png" }
-            .forEach { referenceFile ->
-                val relativePath = referenceFile.relativeTo(snapshotsDir)
-                val compareFile = compareDir.resolve(relativePath.path)
-
-                if (!compareFile.exists()) {
-                    failures.add("Missing comparison file: ${relativePath.path}")
-                } else {
-                    // Basic file size comparison (pixel-perfect would require image comparison library)
-                    if (referenceFile.length() != compareFile.length()) {
-                        failures.add("Size mismatch: ${relativePath.path} (ref: ${referenceFile.length()} bytes, compare: ${compareFile.length()} bytes)")
-                    }
-                }
-            }
-
-        if (failures.isNotEmpty()) {
-            val report = buildString {
-                appendLine("\n========== Snapshot Comparison Failures ==========")
-                appendLine("Total failures: ${failures.size}")
-                appendLine()
-                failures.forEach { appendLine("  - $it") }
-                appendLine("==================================================")
-            }
-            println(report)
-            throw GradleException("Snapshot comparison failed with ${failures.size} differences. See report above.")
-        } else {
-            println("\n✅ All snapshots match! No differences found.")
-        }
-    }
+    description = "Compares snapshots and generates comparison images"
+    dependsOn("compareRoborazziDebug")
 }
 
 /**
@@ -213,9 +159,10 @@ tasks.register("cleanSnapshots") {
     group = "snapshot"
     description = "Cleans all snapshot artifacts and comparison directories"
     doLast {
-        delete(file("src/test/snapshots"))
-        delete(file("src/test/snapshots-compare"))
-        delete(file("out"))
+        delete(fileTree("src/test/resources/roborazzi") {
+            include("**/*.png")
+        })
+        delete(file("build/outputs/roborazzi"))
         println("Snapshot directories cleaned")
     }
 }
@@ -230,7 +177,7 @@ tasks.register("snapshotReport") {
     group = "snapshot"
     description = "Generates a report of all snapshot tests"
     doLast {
-        val snapshotsDir = file("src/test/snapshots")
+        val snapshotsDir = file("src/test/resources/roborazzi")
 
         if (!snapshotsDir.exists()) {
             println("No snapshots found. Run recordSnapshots first.")
@@ -246,12 +193,15 @@ tasks.register("snapshotReport") {
             appendLine("Total snapshots: ${images.size}")
             appendLine("Snapshot directory: ${snapshotsDir.absolutePath}")
             appendLine()
-            appendLine("Snapshots by category:")
 
-            images.groupBy {
-                it.parentFile.name
-            }.forEach { (category, files) ->
-                appendLine("  $category: ${files.size} snapshots")
+            val byTest = images.groupBy {
+                // Extract test class name from filename
+                it.nameWithoutExtension.substringBefore("_")
+            }
+
+            appendLine("Snapshots by test class:")
+            byTest.forEach { (testClass, files) ->
+                appendLine("  $testClass: ${files.size} snapshots")
             }
             appendLine()
             appendLine("All snapshots:")
