+++ library/src/test/java/co/csadev/colorpicker/snapshots/ComponentSnapshotTests.kt
@@ -8,21 +8,26 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.unit.dp
-import app.cash.paparazzi.DeviceConfig
-import app.cash.paparazzi.Paparazzi
 import co.csadev.colorpicker.compose.AlphaSlider
 import co.csadev.colorpicker.compose.ColorPreviewBox
 import co.csadev.colorpicker.compose.ColorWheel
 import co.csadev.colorpicker.compose.LightnessSlider
 import co.csadev.colorpicker.state.ColorPickerState
+import com.github.takahirom.roborazzi.captureRoboImage
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.annotation.Config
+import org.robolectric.annotation.GraphicsMode
 
 /**
- * Snapshot tests for individual ColorPicker UI components.
+ * Snapshot tests for individual ColorPicker UI components using Roborazzi.
  *
  * These tests capture visual regression snapshots of:
  * - ColorWheel with different wheel types (FLOWER, CIRCLE)
@@ -32,433 +37,416 @@ import org.junit.Test
  * - AlphaSlider in various states
  * - ColorPreviewBox with different colors
  */
+@RunWith(RobolectricTestRunner::class)
+@GraphicsMode(GraphicsMode.Mode.NATIVE)
+@Config(sdk = [33], qualifiers = "w411dp-h891dp-normal-long-notround-any-420dpi-keyshidden-nonav")
 class ComponentSnapshotTests {
 
     @get:Rule
-    val paparazzi = Paparazzi(
-        deviceConfig = DeviceConfig.PIXEL_5,
-        theme = "android:Theme.Material3.Light.NoActionBar",
-        maxPercentDifference = 0.0
-    )
+    val composeTestRule = createComposeRule()
 
     // ========== ColorWheel Tests ==========
 
     @Test
     fun colorWheel_flowerType_default() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.FLOWER,
-                            density = 10,
-                            lightness = 1f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.FLOWER,
+                        density = 10,
+                        lightness = 1f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_circleType_default() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.CIRCLE,
-                            density = 10,
-                            lightness = 1f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.CIRCLE,
+                        density = 10,
+                        lightness = 1f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_flowerType_lowDensity() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.FLOWER,
-                            density = 6,
-                            lightness = 1f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.FLOWER,
+                        density = 6,
+                        lightness = 1f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_flowerType_highDensity() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.FLOWER,
-                            density = 15,
-                            lightness = 1f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.FLOWER,
+                        density = 15,
+                        lightness = 1f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_circleType_lowDensity() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.CIRCLE,
-                            density = 6,
-                            lightness = 1f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.CIRCLE,
+                        density = 6,
+                        lightness = 1f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_circleType_highDensity() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.CIRCLE,
-                            density = 15,
-                            lightness = 1f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.CIRCLE,
+                        density = 15,
+                        lightness = 1f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_flowerType_darkLightness() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.FLOWER,
-                            density = 10,
-                            lightness = 0.3f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.FLOWER,
+                        density = 10,
+                        lightness = 0.3f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_circleType_darkLightness() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.CIRCLE,
-                            density = 10,
-                            lightness = 0.3f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.CIRCLE,
+                        density = 10,
+                        lightness = 0.3f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_flowerType_mediumLightness() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.FLOWER,
-                            density = 10,
-                            lightness = 0.6f,
-                            alpha = 1f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.FLOWER,
+                        density = 10,
+                        lightness = 0.6f,
+                        alpha = 1f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorWheel_flowerType_withTransparency() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
-                        ColorWheel(
-                            wheelType = ColorPickerState.WheelType.FLOWER,
-                            density = 10,
-                            lightness = 1f,
-                            alpha = 0.5f
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(300.dp).padding(16.dp)) {
+                    ColorWheel(
+                        wheelType = ColorPickerState.WheelType.FLOWER,
+                        density = 10,
+                        lightness = 1f,
+                        alpha = 0.5f
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     // ========== Slider Tests ==========
 
     @Test
     fun lightnessSlider_fullBrightness() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        LightnessSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color.Red,
-                            lightness = 1f,
-                            onLightnessChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    LightnessSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color.Red,
+                        lightness = 1f,
+                        onLightnessChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun lightnessSlider_mediumBrightness() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        LightnessSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color.Blue,
-                            lightness = 0.5f,
-                            onLightnessChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    LightnessSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color.Blue,
+                        lightness = 0.5f,
+                        onLightnessChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun lightnessSlider_lowBrightness() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        LightnessSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color.Green,
-                            lightness = 0.2f,
-                            onLightnessChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    LightnessSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color.Green,
+                        lightness = 0.2f,
+                        onLightnessChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun lightnessSlider_differentColors() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        LightnessSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color(0xFFFF9800),
-                            lightness = 0.7f,
-                            onLightnessChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    LightnessSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color(0xFFFF9800),
+                        lightness = 0.7f,
+                        onLightnessChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun alphaSlider_fullOpacity() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        AlphaSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color.Red,
-                            alpha = 1f,
-                            onAlphaChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    AlphaSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color.Red,
+                        alpha = 1f,
+                        onAlphaChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun alphaSlider_halfOpacity() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        AlphaSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color.Blue,
-                            alpha = 0.5f,
-                            onAlphaChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    AlphaSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color.Blue,
+                        alpha = 0.5f,
+                        onAlphaChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun alphaSlider_lowOpacity() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        AlphaSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color.Green,
-                            alpha = 0.2f,
-                            onAlphaChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    AlphaSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color.Green,
+                        alpha = 0.2f,
+                        onAlphaChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun alphaSlider_differentColors() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
-                        AlphaSlider(
-                            modifier = Modifier.fillMaxWidth().height(48.dp),
-                            currentColor = Color.Magenta,
-                            alpha = 0.8f,
-                            onAlphaChange = {}
-                        )
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+                    AlphaSlider(
+                        modifier = Modifier.fillMaxWidth().height(48.dp),
+                        currentColor = Color.Magenta,
+                        alpha = 0.8f,
+                        onAlphaChange = {}
+                    )
                 }
             }
         }
+        captureRoboImage()
     }
 
     // ========== ColorPreviewBox Tests ==========
 
     @Test
     fun colorPreviewBox_red() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
-                        ColorPreviewBox(color = Color.Red)
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
+                    ColorPreviewBox(color = Color.Red)
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorPreviewBox_blue() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
-                        ColorPreviewBox(color = Color.Blue)
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
+                    ColorPreviewBox(color = Color.Blue)
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorPreviewBox_green() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
-                        ColorPreviewBox(color = Color.Green)
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
+                    ColorPreviewBox(color = Color.Green)
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorPreviewBox_customColor() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
-                        ColorPreviewBox(color = Color(0xFF9C27B0))
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
+                    ColorPreviewBox(color = Color(0xFF9C27B0))
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorPreviewBox_transparent() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
-                        ColorPreviewBox(color = Color(0x80FF0000))
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
+                    ColorPreviewBox(color = Color(0x80FF0000))
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorPreviewBox_black() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
-                        ColorPreviewBox(color = Color.Black)
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
+                    ColorPreviewBox(color = Color.Black)
                 }
             }
         }
+        captureRoboImage()
     }
 
     @Test
     fun colorPreviewBox_white() {
-        paparazzi.snapshot {
-            MaterialTheme {
-                Surface {
-                    Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
-                        ColorPreviewBox(color = Color.White)
-                    }
+        composeTestRule.setContent {
+            TestTheme {
+                Box(modifier = Modifier.size(100.dp).padding(16.dp)) {
+                    ColorPreviewBox(color = Color.White)
                 }
             }
+        }
+        captureRoboImage()
+    }
+}
+
+@Composable
+private fun TestTheme(content: @Composable () -> Unit) {
+    MaterialTheme {
+        Surface {
+            content()
         }
     }
 }
